still rewriting nodes and catching
