+++ client/utils/helperfunctions.js
@@ -44,18 +44,42 @@ export const LLNodeArray_ = (nodeArray) => {
         console.log('index', index)
         if (typeof(node) !== 'number' && index === 0){
             node.value = node.value
-            node.next = index+1;
-            node.previous = null;
+            node.nextIdx = index+1;
+            node.previousIdx = null;
         }
         else if (typeof(node) !== 'number') {
             node.value = node.value
-            node.next = index+1;
-            node.previous = index-1;
+            node.nextIdx = index+1;
+            node.previousIdx = index-1;
         }
     })
     nodeArray[nodeArray.length - 1].next = null;
     console.log('array after LLNodeArray', nodeArray)
     return nodeArray;
+}
+
+
+export const CleanArray_ = (nodeArr) => {
+    nodeArr.map((node, index) => {
+        node.next = null;
+        node.previous = null;
+    })
+    return nodeArr;
+}
+
+
+//
+//LinkedList.js
+export const nodeArray_ = (nodes) => {
+    // create an array of linked list from 'nodes' class(object)
+    const nodeArr = []; // initialize empty nodeArr.  important!
+    let oldHead = nodes.head; // this is startpoint
+    // push nodes into arr until there is no more 'oldHead'
+    while (oldHead !== null && oldHead !== undefined) {
+      nodeArr.push(oldHead);
+      oldHead = oldHead.next || null;
+    }
+    return nodeArr
 }
 
 //
